+++ src/test/java/com/leroy/magportal/ui/models/orders/ToGiveAwayProductCardData.java
@@ -10,12 +10,12 @@ public class ToGiveAwayProductCardData {
 
 
     private String reasonForNonGiveaway;
-    private Integer createdQuantity;
-    private Integer orderedQuantity;
-    private Integer collectedQuantity;
-    private Integer givenAwayQuantity;
-    private Integer refundToClient;
-    private Integer toGiveAwayQuantity;
+    private Double createdQuantity;
+    private Double orderedQuantity;
+    private Double collectedQuantity;
+    private Double givenAwayQuantity;
+    private Double refundToClient;
+    private Double toGiveAwayQuantity;
 
 
     public void increaseToGiveAwayQuantity(int val) { this.toGiveAwayQuantity += val;
